backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
firebase-admin==6.2.0
python-dotenv==1.0.0
PyPDF2==3.0.1
Pillow>=10.0.0
ffmpeg-python==0.2.0
requests==2.31.0
werkzeug==2.3.7
